+++ create-app/templates/core/babel.config.js
@@ -15,6 +15,7 @@ module.exports = {
     }],
   ],
   plugins: [
+    '@babel/plugin-transform-runtime',
     '@babel/plugin-syntax-dynamic-import',
   ],
 };