+++ create-app/templates/react/babel.config.js
@@ -16,6 +16,7 @@ module.exports = {
     }],
   ],
   plugins: [
+    '@babel/plugin-transform-runtime',
     '@babel/plugin-syntax-dynamic-import',
   ],
 };