+++ create-app/templates/vue/babel.config.js
@@ -15,7 +15,7 @@ module.exports = {
     }],
   ],
   plugins: [
-    'transform-vue-jsx',
+    '@babel/plugin-transform-runtime',
     '@babel/plugin-syntax-dynamic-import',
   ],
 };